// pbweberogwcl/src/app/revoche/commons/models/dettaglio-revoca.ts
/*
 * Copyright Regione Piemonte - 2024 
 * SPDX-License-Identifier: EUPL-1.2 
*/

/* tslint:disable */
export interface DettaglioRevoca {
	codiceCausaleDisimpegno?: string;
	dataRevoca?: string;
	descPeriodoVisualizzata?: string;
	estremiDeterminaRevoca?: string;
	idAnnoContabile?: string;
	idModalitaRecupero?: number;
	idMotivoRevoca?: number;
	idRevoca?: number;
	importoAgevolato?: number;
	importoErogato?: number;
	importoRecuperato?: number;
	importoRevoca?: number;
	importoRevocato?: number;
	modalitaAgevolazione?: string;
	noteRevoca?: string;
	ordineRecupero?: string;
	dtDecorRevoca?: string;
}